+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 
 X.Y.Z (DD-MM-YYYY)
 ------------------
+* Implement a StructRef ESKernel(:pr:`64`)
 * Move ``sympy`` into ``scripts`` and ``test`` dependency groups (:pr:`65`)
 * Provide distinct {Boolean,Float,Integer,String}DatumLiteral types (:pr:`63`)
 * Datums containing integers, booleans and strings are converted to numba {Integer,Boolean,String}Literal types (:pr:`62`)